Do a little bit of re-organization.
The read-write split is too complex with the id vs url style as well, so
it's in one doc now.  It's sorta good that it's one page because
everything is right there, but it is a little long. It might be better
to change this at some point, but let's keep it to one page per thing
for now.

I also changed the front page to explain what's up in a little bit more
detail.