Introduce means to request related documents.
Describes how a server may optionally support inclusion of documents related to the
primary document either by default or via an `include` parameter specified in the request.

Adds a new `Fetching` section that parallels `Updating`. This section can be expanded
to include other means by which custom requests can be made.

Incidentally, some of the headings in the `Updating` section needed further indentation.

[Closes #19]