Unify `id` requirement-level to SHOULD
The `id` requirement-level was inconsistent between the ID-based and
URL-based formats. Based on a [GitHub discussion][ghd], it seemed like
`id` keys are *usually* a good thing, but there are cases where a hard
requirement would be unnecessary. It seems like the spec has settled on
a simple [recommendation][rec] on the presence of the `id` key.

[ghd]: https://github.com/json-api/json-api/issues/181
[rec]: http://www.ietf.org/rfc/rfc2119.txt