Introduce `supported-ext` media type parameter.
The `supported-ext` media type parameter is used by servers to indicate
support for a media type extension. The actual media type extension
used in a document is still specified with the `ext` media type 
parameter.

Also eliminates repetition of the base specification's extension rules
in the bulk and patch extensions.