reserve attributes in top level of resource object
This walls off implementation specific keys for resource objects
under the namespace `attributes`. With this adjustment, all top
level members of resource objects are reserved by JSON-API.

This greatly improves our ability make additive changes that
fit with the current style of the specification in a post 1.0
world.

This proposal was met with wide acceptance by implementors in
both #573 and #588.