Introduce "resource identifier objects".
Instead of referring to "linkage objects" that identify resources with
`type` and `id`, this concept is generalized with a new definition:
"resource identifier objects".

Using this term throughout the spec provides simplification and
consistency, as it can be used not only for relationships but also
for bulk deletes.

Furthermore, with the introduction of the relationships object, we no
longer need to force the term "linkage" into the document structure.
Within a relationships object, this member has been replaced simply with
`"data"`, which represents the relationship data. This has a nice
parallel for relationship URLs in which relationship data _is_ the
primary data.