Reorganize Resource Object Section
- Merge “Attributes Object” and “Attributes” section into one, for
parallelism with the treatment of relationships, which doesn’t start
with a special section describing the “relationships object”. This also
makes the hierarchy simpler.

- Put the Resource Identification section above the Attributes section,
to match the structure of the example payload and the opening of the
“Resource Objects” section.

- But merge the “Resource Types” and “Resource IDs” section into the
“Resource Identification” section, to simplify the hierarchy and make
the whole block much more compact so readers don’t get bogged down with
it. Also, remove some redundancies. For example, the text before said:
“Every resource object is uniquely identified by the combination of its
`"type"`
and `"id"` members. A resource object's `"type"` and `"id"` pair
**MUST** refer to a single, unique resource.”, but these two sentences
both said exactly the same thing.

- Put the “Fields” section right after the attributes and relationship
sections where it clearly belongs, rather than after the Links section.