Put note about extensibility upfront, but move details to the bottom
The extensions section doesn’t need to be upfront, since the details
are irrelevant to most users. All that needs to be upfront is the idea
that the spec can be extended as necessary. So this commit adds a
message to that effect at the top, and then puts the extending section
at the end.