JSON API reaches final review phase.
Due to the massive number of changes to the specification in the past
two weeks, we are providing an extra week for final review before
tagging 1.0.

This release candidate, technically RC4, exists solely to provide time
for the specification to be carefully reviewed for inconsistencies and
defects. Do not expect any changes to the specified document structure.

Thanks to the supportive (and prolific) group of contributors who have
helped us reach this milestone. Let's ensure that the final spec is
as good as can be!