Narrower sidebar
Right now, the sidebar looks much narrower visually than it actually
is, because the text just doesn’t happen to be going up to the right
edge of the <ol> on the most pages. Therefore, it looks like the gutter
between the sidebar and the body content is wider than it really is.
However, if we try to add the sidebar on, e.g., the FAQ page (which I’d
actually like to do), it becomes clear that there needs to be more of a
gutter between the body text. This commit accomplishes that, adding
20px of gutter (10px by narrowing the sidebar; 10px by narrowing the
main column).