Don't hide sidebar until it's necessary
The sidebar is helpful and the text measure is a bit too long anyway on
big screens, so keep the sidebar around until we really have to ditch
it (i.e. after we’ve shrunk the text width a bit and even shrunk the
gutter between the text and the sidebar).

Also adds back the `overflow-y` auto, which got lost in the last merge.