Demonstrate relationships in included resource objects
Even though the spec text says that the objects in `included` are
resource objects (and so can have relationships), I’ve had a couple
people tell me they weren’t sure whether this was allowed because we
didn’t have an example showing it. This aims to rectify that.

Even though the spec text says that the objects in `included` are
resource objects (and so can have relationships), I’ve had a couple
people tell me they weren’t sure whether this was allowed because we
didn’t have an example showing it. This aims to rectify that.