Convert format page to a "format" collection
This will let us add all the different versions of the format as files
in the same Jekyll collection. By grouping them in a collection, we’re
more easily able to loop over just the format pages and add custom
properties to all of them (like that they should all show the sidebar)
without repeating ourselves.

Note that using collections also required updating jekyll, which has
the side compatibility benefit of bringing us inline with the version
of jekyll that github pages is using.