Add new version routes
This creates the following URIs:

/format — always shows the latest version

/format/upcoming — shows a draft of the next version of the spec. I
think this is useful to have on the site so that potential contributors
can review the upcoming changes more easily than if they were tucked
away in a branch or PR somewhere.

/format/X.Y where X.Y is the version number (e.g. /format/1.0) — holds
a copy of version X.Y of the spec. This will be an archived, unchanging
copy (labeled as such) if/once that version of the spec has already been
finalized and a changing draft of version X.Y until that version is
finalized (also labeled as such).

All the logic for labeling the spec pages with appropriate titles (i.e.
calculating what’s a draft, what’s current, etc) is handled by the
title_for_version include, which means that all contributors have to do
is submit text to the index.md file in the appropriate version’s
folder. Then, when we release a new version, we just bump the
`latest_version` setting in _config.yml.