Fix gh-pages build issue
`highlighter: null` is a valid jekyll option, and it's the jekyll 2.4 equivalent of the value we had before (`pygments: false`). And yet, gh pages rejects it with a build warning, saying that it'll automatically use `highlighter: pygments` instead. So this commit just puts in that value, silencing Github's complaint, but not actually changing any of the highlighting behavior (which looks identical on my machine regardless of the value used here).