Redirect from old status page
This page was obsoleted by #865 and is no longer linked to from
anywhere, so this commit replaces it with a simple redirect.

The changes to the pygments setting in _config.yml reflects changes to
Jekyll in 2.4. The removal of the watch flag in the rakefile just a
convenience: watch is already on by default in the version of jekyll
we’re now using!