Redefine 204 Status Code conditions for POST requests
PR for  issue #945:

For POST request the 204 status code depends on if a client-generated ID has been used or not, because in this case the ID isn't generated by the server and so - theoretically - the resulting resource is the same as the resource in the request - BUT this isn't true in all cases and so a 204 status can be wrong sometimes.

Problem 1:
For PUTs and DELETEs, this also depends on added meta data. This can be also added to POST requests, but isn't handled there. So meta data seems to be treated as important data for the client in other cases, but not in case of POSTs.

Problem 2:
The server could use default values or calculate values (the derived fields topic) when creating resources, so the client doesn't have to (or is not allowed to) set all fields in the request. So when the client generated the ID and a 204 response was returned, the client would not have the up-to-date/complete resource data.