Error object: add `type` link
This enables the client to distinguish between different error
conditions that share the same status code, by matching on the `type`
URI.

This is needed for the extension system, where the client needs to
distinguish between the `profile` parameter not being supported by
server, vs. the client providing insufficient profile extensions, both
of which trigger a 415.

This is the start of what I suggested in #683.