Add the concept of "mappings"
This key will be used, for now, to assign a simple/pretty name to each
used extension (identified by its longer URI).

The key is called “mappings”, rather than “aliases” because, in the
future, I suspect that not every key in this object will be an alias
for (i.e. directly substitutable with) its value. Instead, I can
imagine this object also holding, e.g., a mapping between relationship
names and URIs that can be used as an extended link relation to capture
each relationship.