Cleanup `code` formatting
Make the `code` inherit its weight by default, so it can be used inline
with bold text, but set its weight to 600  in particular when its
within an h5 (since h5’s weight of 500 happens to be normal weight
still for common monospace fonts, which only have two weights).

Remove the `.highlight pre` font-family declaration, since it was never
getting applied (the `.highlight pre` always contains a `code`, and
that’s the font that was being used).