Fix broken page titles
We generate page titles in a special way for pages in the _format
collection. Before, we were determining whether a page was a
specification page (i.e., in the _format collection) by looking at
whether it had an empty `page.title`. But, at some point, probably with
Jekyll 3, all Jekyll started assigning all the pages in the format
collection a default page.title of “Index”, based on their file name,
so “Index” was showing up on the site as the title on all specification
pages. Now, we detect specification pages more reliably by reading
page.is_spec_page, which is a variable automatically set to true by
config.yml just for files in the format collection.