Fix duplicate range in character class (in schema)
This removes redundancy in a couple character classes within regular
expressions in the JSON schema. Specifically, there were 2 places where
a character class like `[\w_]` was being used, which is redundant
because `\w` already includes `_`.

I happened to notice this because I was using the `json-schema-rspec`
library in a Ruby project (with warnings turned on), and saw the
following printed to the console:

  warning: character class has duplicated range: /^(?!relationships$|links$)\w[-\w_]*$/