Cleanup `code` formatting
Make the `code` inherit its weight by default, so it will be bold when used as part of a line of bold text (e.g., in a heading). This strategy doesn't work when the `code` is in an h5, since h5’s weight of 500 happens to be a normal weight still for common monospace fonts, so set its weight to 600 explicitly when it's within an h5.