Error object: add type, missing; remove code
1. We add a type member to enables the client to distinguish between different error conditions that share the same status code, by matching on the `type`
URI.

This is needed for the extension system, where the client needs to distinguish between the `profile` parameter not being supported by server, vs. the client providing insufficient profile extensions, both of which trigger a 415.

However, its also a huge win for developer ergonomics, as having an error URL you can follow to get more infomration about a problem is convenient and common.

This is the start of what I suggested in #683.

2. We add the `missing` member as a way for the server to indicate that a required extension is missing. However, we also use this member to clean up the definition of the pointer and parameter fields (as it's not clear that they could previously be used for *missing* document values/parameters).

3. We deprecate the `code` member, because it's functionality is better served by the type field. In particular, application-specific values can still be used (just by
replacing `code` with a `type` URI that the server controls), but `type` also allows the reuse of well-known error types—both those defined in the JSON API spec (which will be necessary for extensions) and those used by, e.g., problem+json.

Removing members from the definition of an error object is NOT a breaking change. Old APIs that send these members will still work with old clients, and new clients won’t choke either, as they’ll simply
ignore the old (and no-longer-defined) members. And, because these values were never required, all clients had to be able to function without them anyway. [See also my inline not about allowing clients to
continue to support old members, if they want.] However, making it illegal to produce certain error object members going forward has the advantage of simplifying an error format that’s redundant right now and a bit underspecified.