Introduce Profiles to v1.1 (#1268)
* Introduce Profiles to v1.1

Profiles are a means to describe additional semantics the JSON API media type, without altering the basic semantics.

Profiles are identified by URI, which ideally should dereference to a document that describes the semantics of the profile.

One or more profiles may be associated with the JSON API media type through the `profile` media type parameter. The application of profiles to a particular document can be specified by clients and servers  via the `Content-Type` header. The application of one or more profiles to a response can be requested by a client via the `Accept` header.

In order to require the application of one or more profiles, the profile(s) must be specified with the `profile` query parameter.

Any structural elements introduced by a profile can be aliased. This allows for better adaptability and composition of profiles. A new profile descriptor object is introduced which allows for declarations of aliases in a particular document.

Closes #1207

* profiles: tweak definitions a bit

E.g., clarify that profiles can cover any implementation-specific query
params, not just sort/filter/page

* profiles: add should about extensibility

* profiles: add note about how to serialize space-separated media type param values

* profiles: make it a MUST to list the applied profiles in Content-Type

* profiles: clarify role of profile in `Accept` per prior convos

* profiles: clarify that `Accept` may be totally ignored

This is key for servers that don’t want to add `Vary: Accept` for
better caching

* profiles: add example of using Accept to request profile

* profiles: clarify that keywords must be valid member names

Should be obvious, but worth saying

* profiles: clarify evolution requriments

The link to the w3c tag terminology about backwards and forwards
compatibility is more precise than “evolve additively”, though it
doesn’t change the intended meaning

* profiles: add note about 415 handling from 1.0 servers

From 1195

* Add top-level links.profile

* Fix typo and trailing whitespace

* Use the term `profile` in example (instead of `ext`)

* Clarify that links can specify URIs, not just URLs

* Replace all usages of jsonapi/profiles with links/profile

Also add examples of profiles

* Profiles: fix broken links

* profiles: remove overloaded uses of term "alias"

* profiles: editorial/terminology cleanup/tweaks

* profiles: define what a profile is

* profiles: add must-ignore rule to example timestamps profile

Technically, this rule is required to make it legal to add even a new,
optional member to the timestamps object.

* profiles: create separate authoring section

This commit just pulls the existing guidelines about authoring profiles
into a dedicated section, without changing any of the substance.

I’m pulling this out because there are a few requirements for authoring
profiles that I think are missing, and it was getting a bit messy to
have the rules for authoring profiles all mixed in with the rules for
using them.

* profiles: add new minimal extra authoring rules

* profiles: editorial tweaks

* profiles: define handling of unrecognized values

* proflies: document extensibility concerns/requirements

* profile: define implicit profile query param value

* profiles: clarify shallow definition of elements

* profiles: small cleanup

* profiles: fix profile param value inference

Whoops.