docs: format page reform
The `/format/upcoming/` url still exists (in case it’s ever necessary
to reference “the next version, whatever that may be at the time the
link is followed”). However, it now simply redirects to `/format/1.1/`
rather than having its own content. It’s status content previously was
a bit confusing, and there wasn’t much reason for it to be a dedicated
page.

Additionally, it was confusing to have two links to v1.1 in the version
picker (the upcoming link and the 1.1 permalink). So, now, the version
picker excludes a separate permalink entry for v1.1, and instead has
the picker entry for the “Upcoming spec” simply go to the
`/format/1.1/` page.