Remove redirect in link to PATCH RFC
Fixes #1588

In the FAQ, Where's PUT?, the link under PATCH points to an insecure
location, e.g., http://tools.ietf.org/html/rfc5789 and results in
two (2) redirects to the correct location.

The correct URL, after redirection, is
https://datatracker.ietf.org/doc/html/rfc5789

Moreover, this can sometimes result in a 404 from tools.ietf.org
(which appears to be intermittent, possibly due to a server
misconfiguration; the personal website for Henrik Levkowetz, and
member of the IETF Tools Team, appeared at the root several times
before starting to redirect, see response in issue 🤔 ).