only client-requested sparse fieldsets are exception to full linkage requirement
The current wording could be misunderstood that also fields excluded by the server if client does _not_ request a sparse fieldset are an exception for full linkage requirement. This changes clarifies that only exception is if a client _explicilty_ requested a sparse fieldset, which does not include a relationship required for full linkage.