Clarify processing rules for malformed `ext` and `profile` members of JSON:API object (#1607)
* Clarify processing rules for malformed `ext` and `profile` members of JSON:API object

The applied extension and profiles must be provided as `ext` and `profile` media type parameters of `Content-Type` header. Additionally they may be listed as `jsonapi.ext` and `jsonapi.profile` member of a JSON:API document.

This opens up the possibility of having different values in both places. So far the specification has not defined processing rules for such malformed documents.

This clarifies that a client and server must honor the `Content-Type` header in case of a conflict. Additionally it clarifies that they should not reject the document but ignore the invalid `jsonapi.ext` and/or `jsonapi.profile` members.

Ignoring them in case of a conflict rather than rejecting the document allows server and clients to _not_ process them at all - even if present. This avoids computing overhead just for the sake of validating the document.

* Update _format/1.1/index.md

Co-authored-by: Jeldrik Hanschke <jelhan@users.noreply.github.com>

* Update _format/1.1/index.md

Co-authored-by: Dan Gebhardt <dan@cerebris.com>