Standardize and clarify success (2xx) responses
Uses consistent language across requests that create, update, and delete resources and relationships.

Re-orders success codes to be in ascending order for consistency.

Clarifies that it is always ok to return a 200 (or 201 in the case of creation) response with a document that does not include primary data when the server does not change the targeted resource(s) / relationship(s) in any way different from what was requested. This allows the server to include other top-level members, such as `meta`, in responses without primary data.